Rename to NetSuiteSOAP as we intend to make the REST API adaptor our primary netsuite adaptor
