+++ lib/ledger_sync/adaptors/netsuite_soap/config.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+LedgerSync.register_adaptor(:netsuite_soap, module_string: 'NetSuiteSOAP') do |config|
+  config.name = 'NetSuite SOAP'
+end
