+++ lib/ledger_sync/adaptors/netsuite_soap/operation/sparse_update.rb
@@ -4,19 +4,17 @@ require_relative '../operation'
 
 module LedgerSync
   module Adaptors
-    module NetSuite
+    module NetSuiteSOAP
       module Operation
-        class Find
-          include NetSuite::Operation::Mixin
+        class SparseUpdate
+          include NetSuiteSOAP::Operation::Mixin
 
           private
 
           def operate
-            return failure(nil) if resource.ledger_id.nil?
-
-            response = adaptor.find(
+            response = adaptor.post(
               resource: netsuite_online_resource_type,
-              id: resource.ledger_id
+              payload: merged_serializer.to_ledger_hash
             )
 
             success(
