+++ lib/ledger_sync/adaptors/netsuite_soap/subsidiary/operations/create.rb
@@ -2,25 +2,29 @@
 
 module LedgerSync
   module Adaptors
-    module NetSuite
+    module NetSuiteSOAP
       module Subsidiary
         module Operations
-          class Find < NetSuite::Operation::Find
+          class Create < NetSuiteSOAP::Operation::Create
             class Contract < LedgerSync::Adaptors::Contract
               params do
-                required(:external_id).maybe(:string)
-                required(:ledger_id).filled(:string)
-                required(:name).maybe(:string)
-                required(:state).maybe(:string)
+                required(:external_id).filled(:string)
+                required(:ledger_id).value(:nil)
+                required(:name).filled(:string)
+                required(:state).filled(:string)
               end
             end
 
             private
 
             def operate
-              netsuite_resource = ::NetSuite::Records::Subsidiary.get(
-                internal_id: resource.ledger_id
+              netsuite_resource = ::NetSuite::Records::Subsidiary.new(
+                external_id: resource.external_id,
+                name: resource.name,
+                state: resource.state
               )
+
+              return netsuite_failure(netsuite_resource: netsuite_resource) unless netsuite_resource.add
 
               resource.name = netsuite_resource.name
               resource.ledger_id = netsuite_resource.internal_id
