--- spec/integrations/netsuite/adaptor_spec.rb
@@ -1,9 +1,0 @@
-require 'spec_helper'
-
-support :netsuite_helpers
-
-RSpec.describe LedgerSync::Adaptors::NetSuite::Adaptor, type: :feature do
-  include NetSuiteHelpers
-
-  let(:adaptor) { netsuite_adaptor }
-end
