+++ spec/spec_helper.rb
@@ -28,7 +28,7 @@ VCR.configure do |config|
   config.hook_into :webmock
   config.configure_rspec_metadata!
 
-  # These need to match the defaults in support/netsuite_helpers.rb for the
+  # These need to match the defaults in support/netsuite_soap_helpers.rb for the
   # stubs to work.
   config.filter_sensitive_data('netsuite_account_id') { ENV['NETSUITE_ACCOUNT_ID'] } if ENV.key?('NETSUITE_ACCOUNT_ID')
   config.filter_sensitive_data('NETSUITE_CONSUMER_KEY') { ENV['NETSUITE_CONSUMER_KEY'] } if ENV.key?('NETSUITE_CONSUMER_KEY')
